+++ lib/dealog_backoffice/messages/events/message_deleted.ex
@@ -1,0 +1,9 @@
+defmodule DealogBackoffice.Messages.Events.MessageDeleted do
+  @derive Jason.Encoder
+  defstruct [
+    :message_id,
+    :title,
+    :body,
+    :status
+  ]
+end