+++ lib/dealog_backoffice/messages/projections/deleted_message.ex
@@ -1,4 +1,4 @@
-defmodule DealogBackoffice.Messages.Projections.MessageForApproval do
+defmodule DealogBackoffice.Messages.Projections.DeletedMessage do
   use Ecto.Schema
 
   alias DealogBackoffice.Ecto.Type.Status
@@ -6,12 +6,10 @@ defmodule DealogBackoffice.Messages.Projections.MessageForApproval do
   @primary_key {:id, :binary_id, autogenerate: false}
   @timestamps_opts [type: :utc_datetime_usec]
 
-  schema "message_approvals" do
+  schema "deleted_messages" do
     field :title, :string
     field :body, :string
     field :status, Status
-    field :reason, :string
-    field :note, :string
 
     timestamps()
   end