+++ lib/dealog_backoffice/messages/projections/message.ex
@@ -1,13 +1,15 @@
 defmodule DealogBackoffice.Messages.Projections.Message do
   use Ecto.Schema
 
+  alias DealogBackoffice.Ecto.Type.Status
+
   @primary_key {:id, :binary_id, autogenerate: false}
   @timestamps_opts [type: :utc_datetime_usec]
 
   schema "messages" do
     field :title, :string
     field :body, :string
-    field :status, :string
+    field :status, Status
     field :rejection_reason, :string
 
     timestamps()